+++ src/AppRouter.jsx
@@ -15,6 +15,7 @@ import VPN from './vpnconfig.jsx'
 import Logout from './logout.jsx'
 import UserManagement from './userManagement.jsx'
 import PPPPage from './ppp.jsx'
+import CameraSwitcherPage from './cameraswitcher.jsx'
 
 function AppRouter () {
   const [isAuthenticated, setIsAuthenticated] = useState(null)
@@ -74,6 +75,7 @@ function AppRouter () {
           <Link className='list-group-item list-group-item-action bg-light' to="/adhoc">Adhoc Wifi Config</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/apclients">Access Point Clients</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/video">Photo and Video</Link>
+          <Link className='list-group-item list-group-item-action bg-light' to="/cameraswitcher">Camera Switcher</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/cloud">Cloud Upload</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/vpn">VPN Config</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/about">About</Link>
@@ -93,6 +95,7 @@ function AppRouter () {
             <Route exact path="/network" element={<NetworkConfig />} />
             <Route exact path="/about" element={<About />} />
             <Route exact path="/video" element={<VideoPage />} />
+            <Route exact path="/cameraswitcher" element={<CameraSwitcherPage />} />
             <Route exact path="/flightlogs" element={<LogBrowser />} />
             <Route exact path="/apclients" element={<NetworkClients />} />
             <Route exact path="/ntrip" element={<NTRIPPage />} />